Don't call chippers adjacent to the data being retrieved.
With a multi-segment NITF, each image segment has its own chipper
and we need to determine which ones to ask for data.  Previously
for a read request that ended at the border of an image segment,
the next image segment was also included in the list of chippers
to call with a request size of 0 rows.  This tripped up the
BIPChipper's file backed access methods, but worked for memory
mapped access.

Now don't include this adjacent image segment/chipper in the list
of places to check for data to avoid this problem entirely.